backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.136.1",
    "jinja2>=3.1.6",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.2.0",
    "pydantic>=2.13.3",
    "pymupdf>=1.27.2.3",
    "pypdf>=6.10.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
    "weasyprint>=68.1",
]
